描画処理を Painter に分離
